requirements.txt
tensorflow
streamlit
opencv-python
matplotlib
numpy
Pillow
gspread
oauth2client
